[enh] Make sound effects generic
