+++ .env.sample
@@ -21,3 +21,4 @@ PAGLOPS_URL=[URL]
 PAGLOPS_USER=[USERNAME]#[DISCRIMINATOR]
 
 BAN_URL=[URL]
+DRUM_URL=[URL]
